backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
chromadb>=0.4.0
python-multipart>=0.0.6
httpx>=0.25.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
chardet>=5.0.0
